+++ phylodb/src/main/java/pt/ist/meic/phylodb/utils/service/Reference.java
@@ -2,7 +2,7 @@ package pt.ist.meic.phylodb.utils.service;
 
 public class Reference<T> extends Entity<T> {
 
-	public Reference(T id, int version, boolean deprecated) {
+	public Reference(T id, Long version, boolean deprecated) {
 		super(id, version, deprecated);
 		this.id = id;
 	}
